requirements.txt
anthropic
streamlit
pymupdf
python-dotenv
sentence-transformers
